Issue #1692, moved logic to another class, and used it in both add_pubkey and contacts files
