+++ extension/js/common/ui/fetch_key_ui.ts
@@ -1,0 +1,37 @@
+/* © 2016-2018 FlowCrypt Limited. Limitations apply. Contact human@flowcrypt.com */
+
+'use-strict';
+
+import { Ui } from '../browser.js';
+import { Catch } from '../platform/catch.js';
+import { KeyImportUi } from './key_import_ui.js';
+import { BrowserMsg } from '../extension.js';
+
+export class FetchKeyUI {
+    handleOnPaste(elem: JQuery<HTMLElement>) {
+        elem.on('paste', Ui.event.handle(async (elem: HTMLInputElement, event) => {
+            const clipboardEvent = event.originalEvent as ClipboardEvent;
+            if (clipboardEvent.clipboardData) {
+                const possiblyURL = clipboardEvent.clipboardData.getData('text');
+                const pattern = /(ftp|http|https):\/\/(\w+:{0,1}\w*@)?(\S+)(:[0-9]+)?(\/|\/([\w#:.?+=&%@!\-\/]))?/;
+                if (pattern.test(possiblyURL)) {
+                    const pubkey = await this.fetchPubkey(possiblyURL);
+                    if (pubkey) {
+                        elem.value = pubkey;
+                    }
+                }
+            }
+        }));
+    }
+
+    private async fetchPubkey(url: string) {
+        try {
+            // tslint:disable-next-line: no-direct-ajax
+            const result = (await BrowserMsg.send.bg.await.ajax({ req: { url, type: 'GET', dataType: 'text', async: true }, stack: Catch.stackTrace() })) as string;
+            const keyImportUi = new KeyImportUi({ checkEncryption: true });
+            return await keyImportUi.checkPub(result);
+        } catch (e) {
+            return;
+        }
+    }
+}
